Rename with new uppercase standard
