+++ cpp/brain/Conversion.cpp
